fix the float number precision bug: change to Decimal

+++ pipeline/ephys.py
@@ -215,7 +215,7 @@ class TrialEvent(dj.Imported):
     trial_event_id: smallint
     ---
     -> experiment.TrialEventType
-    trial_event_time : float  # (s) from session start (global time)
+    trial_event_time : Decimal(10, 5)  # (s) from session start (global time)
     """    
 
 
@@ -229,7 +229,7 @@ class ActionEvent(dj.Imported):
     action_event_id: smallint
     ---
     -> experiment.ActionEventType
-    action_event_time : float  # (s) from session start (global time)
+    action_event_time : Decimal(10, 5)  # (s) from session start (global time)
     """
 
 @schema
